# lakehoused/lakehoused/api/openapi.yaml
openapi: 3.1.0
info:
  title: Amplifier Daemon API
  version: 1.0.0
  description: |
    REST API for the Amplifier daemon - exposes amplifier CLI functionality as HTTP endpoints.
    
    ## Architecture
    - File-based persistence (JSON)
    - SSE for streaming responses
    - Multi-session concurrency support
    - Stateless session resumption
    
    ## Key Concepts
    - **Session**: A persistent conversation context with Claude
    - **Profile**: Agent configuration with system prompt and tools
    - **Execution**: Single-shot or streaming LLM interaction
    
  contact:
    name: Amplifier Project
    url: https://github.com/microsoft/amplifier
servers:
  - url: http://localhost:8080/api/v1
    description: Local development server
paths:
...
  /sessions/{session_id}:
    get:
      summary: Get session details
      description: Retrieve complete session state including conversation history
      operationId: getSession
      tags: [Sessions]
      parameters:
        - $ref: '#/components/parameters/SessionId'
      responses:
        '200':
          description: Session details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SessionDetail'
        '404':
          $ref: '#/components/responses/NotFound'

    delete:
      summary: Delete session
      description: Permanently delete a session
      operationId: deleteSession
      tags: [Sessions]
      parameters:
        - $ref: '#/components/parameters/SessionId'
      responses:
        '204':
          description: Session deleted
        '404':
          $ref: '#/components/responses/NotFound'
...
components:
  parameters:
    SessionId:
      name: session_id
      in: path
      required: true
      description: Unique session identifier (UUID)
      schema:
        type: string
        format: uuid
      example: 123e4567-e89b-12d3-a456-426614174000
  schemas:
...
    SessionDetail:
      type: object
      required: [metadata, messages]
      properties:
        metadata:
          $ref: '#/components/schemas/SessionMetadata'
        messages:
          type: array
          items:
            $ref: '#/components/schemas/Message'
        context:
          type: object
          additionalProperties: true
          description: Additional session context
        config:
          type: object
          additionalProperties: true
          description: Session configuration
    SessionMetadata:
      type: object
      required: [session_id, name, created_at, updated_at, turns]
      properties:
        session_id:
          type: string
          format: uuid
        name:
          type: string
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        turns:
          type: integer
        total_tokens:
          type: integer
        cost_usd:
          type: number
          format: float
        duration_seconds:
          type: number
          format: float
        tags:
          type: array
          items:
            type: string
    Message:
      type: object
      required: [role, content, timestamp]
      properties:
        role:
          type: string
          enum: [user, assistant, system]
        content:
          type: string
        timestamp:
          type: string
          format: date-time
        metadata:
          type: object
          additionalProperties: true

    # Execution Models
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
              description: Machine-readable error code
            message:
              type: string
              description: Human-readable error message
            details:
              type: object
              additionalProperties: true
              description: Additional error context
  responses:
...
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error:
              code: NOT_FOUND
              message: Session not found
              details:
                session_id: 123e4567-e89b-12d3-a456-426614174000
...